댓글 모델 만들고 댓글 추가

+++ posts/views.py
@@ -1,6 +1,6 @@
 from django.shortcuts import render, redirect, get_object_or_404
-from .forms import PostModelForm
-from .models import Post
+from .forms import PostModelForm, CommentForm
+from .models import Post, Comment
 
 
 # Create your views here.
@@ -30,13 +30,20 @@ def create(request):
 def list(request):
     # 모든 Post를 보여줌.
     posts = Post.objects.all()
+    
+    comment_form = CommentForm()
     context = {
         'posts':posts,
+        'comment_form':comment_form,
     }
     return render(request,'posts/list.html',context)
     
 def delete(request, post_id):
     post = Post.get_object_or_404.get(Post,pk=post_id)
+    
+    if post.user != request.user:
+        return redircet('posts:list')
+    
     post.delete()
     return redirect('posts:list')
     
@@ -63,4 +70,16 @@ def update(request, post_id):
             'form' : form,
         }
         return render(request, 'posts/update.html', context)
+    
+def create_comments(request, post_id):
+    comment_form = CommentForm(request.POST)
+    
+    if comment_form.is_valid():
+        comment = comment_form.save(commit=False) # 일단 
+        comment.user = request.user # 누가 작성했는지 넣어주고
+        comment.post_id = post_id
+        comment.save()
+        return redirect('posts:list')
+        
+
